refactor: auth파일에 password-reset관련 저장
