+++ src/main/java/com/example/kspot/users/controller/UsersController.java
@@ -145,10 +145,4 @@ public class UsersController {
                 .body(new ApiResponseDto<>(200, "로그인 성공", token.accessToken()));
     }
 
-    @PatchMapping("/reset-password")
-    public ResponseEntity<ApiResponseDto<?>> resetPassword(@RequestBody UserRequestDto dto) {
-        userService.resetPassword(dto);
-        return ResponseEntity.ok(new ApiResponseDto<>(200 , "비밀번호 변경에 성공했습니다!" , null));
-    }
-
 }
