+++ src/test/java/com/example/kspot/users/UserServiceTest.java
@@ -2,11 +2,9 @@ package com.example.kspot.users;
 
 import com.example.kspot.config.SecurityConfig;
 import com.example.kspot.email.service.EmailVerificationService;
-import com.example.kspot.jwt.JwtProvider;
+import com.example.kspot.auth.jwt.JwtProvider;
 import com.example.kspot.users.dto.UserRequestDto;
 import com.example.kspot.users.dto.UserResponseDto;
-import com.example.kspot.users.dto.UserUpdataRequestDto;
-import com.example.kspot.users.dto.UserUpdateResponseDto;
 import com.example.kspot.users.entity.Users;
 import com.example.kspot.users.exception.NotEmailVerifiedException;
 import com.example.kspot.users.repository.UserRepository;
